requirements.txt
fastapi
uvicorn
python-multipart

opencv-python
numpy
pillow

deepface
faiss-cpu

tensorflow
keras

mtcnn
retina-face

requests
pandas
scikit-learn

python-dotenv
langchain
openai
n8n
requests